Actually compare the contents of ArrayBuffers.
Before, only the properties (of which there are none) were compared.

Also, document.write() is evil. In Firefox, which implements the
specification correctly, it creates a new global object, and a new
hierarchy of objects (so Object before != Object after). This breaks a
lot of things, for example `instanceof`. This is different from other
decent browsers ( WebKit, a fork of WebKit, and possibly Edge), which
don't follow the specification.